comboctl-base: Move ComboPacket into TransportLayer and rename it to Packet
ComboPacket has always been tightly coupled with TransportLayer. As
preparation for upcoming application layer packets and to better reflect
that tight coupling, move ComboPacket into TransportLayer as a subclass.
This also makes it clearer that this is a transport layer packet in the
code, since the type is then "TransportLayer.Packet".

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>