comboctl-base: Use new Twofish implementation written in Kotlin
This is an adaptation of the Java implementation, but written purely in
Kotlin. This is part of an effort to make ComboCtl JVM independent.

The Twofish.kt code has almost 1:1 the same structure as the Java code,
but is smaller and cleaner, since substantial amounts of the code in the
Java implementation are extras for logging and debugging.

Also add unit tests to verify that the implementation works correctly.
They use test vectors from Bruce Schneier's Known Answer test dataset.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>