comboctl-base: Handle failures in background receive loops properly
Failures in the background receive loops cannot be simply propagated
like in other code. Instead, they have to be communicated to the rest of
the application via some sort of notification mechanism to let the rest
of the code know that the loop stopped due to error. Implement such
notifications and API additions.

One noteworthy addition is the new ReceiveLoopFailureException class,
which is thrown when one such background receive loop fails for whatever
reason. When there is a failure in a background loop, any ongoing receive
operation will throw this exception, and an optional callback will be
called to notify the application about the failure. Any future IO attempts
will also throw such an exception. This goes on until the connection
is reestablished. (disconnect() must be called manually first.)

This does not apply to performPairing(), since there is no background loop
running after that call finishes.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>