comboctl-base: Rework the persistent store and its error handling
This improves the PersistentPumpStateStore error handling and the way
stores are requested. It also adds code to MainControl to check if a
pump already has a valid store. If so, it is already paired. That way,
MainControl can avoid duplicate pairing attempts. A new class is
introduced (PersistentPumpStateStoreBackend) that provides the necessary
access to get a PersistentPumpStateStore and look up whether or not a
valid store already exists for a given Bluetooth address.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>