comboctl-base: TransportLayerIO: Fix disconnect process
Send the disconnect packet first before trying to shut down the worker.
This will make the Combo terminate the connection, and the blocking
receive() calls in the worker will be interrupted and throw IO exceptions.
These must be ignored during shutdown to avoid false positives. Introduce
the ignoreBackgroundWorkerErrors flag for that purpose.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>