comboctl-base: ApplicationLayerIO: Add response and no-response send calls
Many commands have corresponding response commands. It is important to
first wait for a response to arrive before another command-with-response
can be sent. Add sendPacketWithResponse() for that purpose, which avoids
messing up the request-response sequences by using a coroutine mutex.
These are rather slow, but commands aren't sent all that often.

Also rename sendPacket() to sendPacketNoResponse() for consistency.

This also changes the CMD_PING loop to wait for a CMD_PING_RESPONSE
instead of just ignoring these response packets.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>