comboctl-base: PumpIO: Improve RT button pressing behavior
Up until now, the code simply waited for 200ms and then sent NO_BUTTON,
not making use of button confirmations. It is beneficial however to do so.
In particular, the loop that was used for long button presses is reworked
to wait for these confirmations instead.

Also, an extra 200ms delay is introduced to not overload the Combo's
packet ringbuffer and cause connection errors.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>