comboctl-base: Simplify device discovery
Scanning and connecting to devices simultaneously turns out to not be
very useful in practice, and complicates the API. Change the discovery
mechanism to automatically stop discovery (like the behavior of the old
discovery mechanism when setting the onlyDiscoverOneDevice argument to
true). As a side effect, startEventHandling and stopEventHandling are
not necessary anymore.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>