comboctl: Introduce "main" namespace, migrate Pump & rename MainControl
This new namespace contains the main/primary API for applications. Classes
in this namespace are the front-end of the ComboCtl API. They can use
code from the "base" and "parser" namespaces.

The Pump class, being such a front-end API, is relocated here. MainControl
is also moved here, and renamed to PumpManager, which better reflects
what this class actually does.

Later on, the new scripter will also reside here.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>