comboctl-base: Make sure sendPacketWithResponse calls behave atomically
If the coroutine that runs the sendPacketWithResponse call is cancelled
after the outgoing packet was sent but before the expected response was
received, the connection ends up in an incorrect state, since the response
that should have been received by the now-cancelled sendPacketWithResponse
call is received by a _different_ sendPacketWithResponse call that expected
a _different_ response.

TODO: What if the call _must_ be cancelled right away, for example during
shutdown because of a fatal error?

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>