comboctl-base: Add overallProgress to progress report
overallProgress is a simple Double in the 0.0 - 1.0 range, which is much
simpler to integrate with progress indicator UI elements. This also allows
for ProgressReporter instances to define a custom overallProgress callback
in case the reporter needs one (bolusDeliveryProgressReporter for example).
That way, users do not need to worry about the individual stages of the
progress reporter.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>