comboctl-main: Override background scope dispatcher with the Default one
There is no benefit in running the background worker with any dispatcher
other than the Default one. It is however easy to introduce a bug when
a scope is passed with a dispatcher that for example uses the UI thread.
In such cases, UI freezes would also freeze the background worker.
Consequently, this commit overrides the scope's dispatcher with the
Default one.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>