comboctl-main: Fix alert screen handling
Previously, checkForAlerts() was incorrectly using the first() collector
repeatedly, which ended up immediately using the frame that was stored
in the display frame flow's replay cache. Replace this with a first()
collector call that uses a predicate that looks at the screen and only
returns true when the screen isn't an alert screen anymore.

This also reuses the same alert screen handling in dispatchCommand()
in case an alert appears during command dispatch. (This is unusual;
alert screens typically show up before and after command dispatch.
Still, the code must be able to handle mid-dispatch alert.)

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>