comboctl-base: Various PumpIO improvements
* Replace connectAsync() with suspending connect()
* Remove packetReceiverScope argument, replacing it with an internal scope
* Fix order of operations in connect(); _first_ connect the Bluetooth
  device, _then_ proceed with the handshaking
* Remove reconnect(), since it was not being used
* Add default maxRequests count to getCMDHistoryDelta()
* Add doRestartHeartbeat argument to sendPacketWithResponse()
  and sendPacketWithoutResponse()
* Fix CMD ping loop (_first_ call delay(), _then_ emit PING packet)

The scope that is being used for running the keep-alive loops and the
long RT button press loops are no longer using the sequenced dispatcher,
since they don't need to. The bits that must be run that way are in
TransportLayer.IO, and the application layer send functions in PumpIO are
surrounded by the sendPacketMutex.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>