comboctl-main: Add Pump class
This is the main control interface for users. It implements a large number
of safety checks and automations to common operations like setting a TBR
or delivering a bolus. Apps typically use this class directly for
controlling the Combo.

The main difference to the old Pump class is that the previous version
did not contain the same level of checks and automations.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>