comboctl-main: Use cyclic quantity range 12 for datetime month adjustments
This allows the logic to reach the target month sooner in some cases.
For example, if "2" is shown on screen, and "11" is the target month,
then this allows the logic to decide to just press DOWN 3 times (the
quantity on screen jumps back to 12 if it is at 1 and DOWN is pressed)
instead of having to press UP 8 times (which takes longer to finish).

Also add a TODO for using a cyclic quantity range for the days adjustments.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>