comboctl-base: Modify Aborted stage to better differentiate between causes
"Aborted" can be reached due to the user cancelling an operation, or
because an error occurred etc. Change Aborted to be an open class, and
write subclasses for these specialized cases. That way, code can just
check for Aborted to catch all of these possible Abort stages, or check
for individual subclasses to apply a specific logic (for example, to
display different UI elements when the operation was cancelled vs. when
a timeout occurred).

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>