comboctl-base: Update connect progress reporter in PumpIO.connect()
Do this there instead of in the BluetoothDevice subclasses to provide
feedback to the user that is not as confusing. Otherwise, the user may
see connection attempts repeat in a loop, for example seeing attempts
#1-#5 fail, only to go back to #1 again etc. on Android. By shifting
the updates to PumpIO.connect(), this does not happen anymore. The
internal connection loop in AndroidBluetoothDevice no longer drives
that progress reporter.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>