comboctl-main: Improve alert screen handling speed and efficiency
Previously, checkForAlerts() always switched to the command mode to
send the CMD_READ_ERROR_WARNING_STATUS packet. This is unnecessary if
the pump is currently in remote terminal mode - if an alert happens,
then an alert screen will already be visible. Change the logic to
just directly parse the current screen if the pump is in the remote
terminal mode.

Also, if a warning screen is to be dismissed, press CHECK twice right
away instead of waiting for a new screen before pressing CHECK the
second time.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>