comboctl-main: Unpair unknown pumps right before pairing new pump
Instead of unpairing unknown pumps in setup(), do it in pairWithNewPump().
Doing this in setup() has one problem: Unpairing is inappropriate if
pump states are to be restored from a serialized form of said states, for
example a configuration file. Instead, unpairing unknown pumps before
pairing with a new pump makes a lot more sense, and avoids this problem,
since restoring pump states does not involve pairWithNewPump() calls.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>